+++ css/src/components/radio.css
@@ -7,6 +7,7 @@
     -ms-user-select: none;
     user-select: none;
     align-items: center;
+    cursor: pointer;
 }
 
 .Radio-outerWrapper {
@@ -26,11 +27,16 @@
 
 .Radio-input {
     opacity: 0;
-    cursor: pointer;
     position: absolute;
     height: 100%;
     width: 100%;
     z-index: 2;
+    margin: 0;
+    cursor: pointer;
+}
+
+.Radio-label {
+    cursor: pointer;
 }
 
 .Radio-wrapper {
@@ -48,10 +54,13 @@
     z-index: 1;
 }
 
+.Radio--disabled {
+    pointer-events: none;
+}
+
 .Radio--disabled .Radio-wrapper {
     border: var(--spacing-xs) solid var(--secondary-light);
     background-color: var(--secondary-lightest);
-    pointer-events: none;
 }
 
 .Radio-input:checked~.Radio-wrapper {
@@ -86,14 +95,8 @@
 
 .Radio--disabled .Radio-input:checked~.Radio-wrapper {
     border: var(--spacing-xs) solid var(--primary-lighter);
-    pointer-events: none;
 }
 
 .Radio--disabled .Radio-input:checked~.Radio-wrapper:after {
     background: var(--primary-lighter);
-}
-
-.Radio--disabled .Radio-wrapper--checked {
-    border: var(--spacing-xs) solid var(--primary-lighter);
-    pointer-events: none;
 }
